backend/marker_service/requirements.txt
Flask==2.3.3
marker-pdf>=0.3.0
PyMuPDF>=1.24.0
camelot-py[cv]==0.11.0
pdfplumber==0.11.0
opencv-python==4.9.0.80
pandas>=1.5.0
tabulate>=0.9.0
gunicorn
